create a Universal & Conway Era
